+++ src/galax/dynamics/_src/dynamics/__init__.py
@@ -6,12 +6,9 @@ __all__ = [
     "AbstractDynamicsField",
     "HamiltonianField",
     "NBodyField",
-    # utils
-    "parse_time_specification",
 ]
 
 from .field_base import AbstractDynamicsField
 from .field_hamiltonian import HamiltonianField
 from .field_nbody import NBodyField
-from .parsetime import parse_time_specification
 from .solver import DynamicsSolver
