+++ src/galax/dynamics/integrate.py
@@ -13,13 +13,13 @@ from jaxtyping import install_import_hook
 from galax.setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("galax.dynamics.integrate", RUNTIME_TYPECHECKER):
-    from ._src.dynamics import parse_time_specification
     from ._src.legacy import (
         Integrator,
         InterpolatedPhaseSpaceCoordinate,
         evaluate_orbit,
     )
     from ._src.orbit import PhaseSpaceInterpolation
+    from ._src.parsetime import parse_time_specification
 
 # Cleanup
 del install_import_hook, RUNTIME_TYPECHECKER
