+++ src/galax/dynamics/solve.py
@@ -19,8 +19,9 @@ from galax.setup_package import RUNTIME_TYPECHECKER
 with install_import_hook("galax.dynamics.solve", RUNTIME_TYPECHECKER):
     from diffraxtra import DiffEqSolver, VectorizedDenseInterpolation
 
-    from ._src.dynamics import DynamicsSolver, parse_time_specification
+    from ._src.dynamics import DynamicsSolver
     from ._src.orbit import Orbit, PhaseSpaceInterpolation, compute_orbit
+    from ._src.parsetime import parse_time_specification
     from ._src.solver import AbstractSolver, SolveState
 
 # Cleanup
